add/remove entity

+++ src/world/entity_spec.rs
@@ -3,5 +3,5 @@ pub struct EntitySpec
     pub arch_id:         usize,
     pub chunk_idx:       usize,
     pub idx_in_chunk:    usize,
-    pub current_version: usize,
+    pub current_version: u32,
 }
